Connect to WiFi and MQTT last
Some wierd issues with the ESP-Wroom-32 not be able to connect to the WiFi led me to belive that my decoupeling of my stepper motor drivers was not good enough. It could connect to the WiFi if supplied by USB, but not if the 12V line was hooked up and the ESP was fed through a 5V-buck.

I soldered on alot of caps on all major power bus inputs and it didn't help.

Then i tried disconnecting the motors and it helped. I had my scope hooked up to the 3V3 rail for the ESP and could see 200mV noise on there if the motors was hooked up.

The enable pin for the stepper drivers was enabled during startup, and it created unnecesary current consumtion and noise on the rails during startup.
I simply moved the WiFi connect last in startup instead of first.
Having the manual override of the steppers is far more important than beeing connect to the MQTT server anyway.

The initialization of WiFi and connect to an access point seamed to be really sensitive for noise.
However, after the WiFi connection is established I could enable the stepper driver without any issues.